requirements.txt
mongoengine
iptools
ipaddress
urlnorm
celery
idna
requests>2.10.0
lxml
dnspython
flask-classy
flask-negotiation
flask-mongoengine
flask-login
tldextract
git+https://github.com/joepie91/python-whois#egg=pythonwhois
ipwhois
blinker
pyyaml
simplejson
yara-python
redis
sphinx
sphinx_rtd_theme
sphinxcontrib-httpdomain
geoip2
psutil
shodan
python-magic
pdfkit
pdfminer
html2text
readability-lxml
beautifulsoup4
requests[socks]
